+++ extension/Changelog.md
@@ -1,5 +1,15 @@
 # BCT Changelog
 
+## 0.6.6
+### Bugfixes:
+- Updated to Item permissions for R105 (By Rama)
+- Tailwag now works in ChatRooms
+
+## 0.6.5
+### Features:
+- Show Room Slots added
+- Friend list would now have an extra column (Slots) showing number of people in your friend's room and max capacity of the room
+
 ## 0.6.4
 ### Bugfixes:
 - Updated for R104